requirements.txt
Flask==2.0.1
Werkzeug==2.0.1
Pillow==8.4.0
opencv-python==4.5.3.56
numpy==1.21.2
pandas==1.3.3
python-dotenv==0.19.0
groq==0.4.2
easyocr==1.6.2
PyMuPDF==1.19.1
python-docx==0.8.11
requests==2.26.0
python-magic==0.4.24
xlsxwriter==3.0.3
gunicorn==20.1.0